Feedback of a restaurant using an Android app
A Feedback Application for a restaurant which helps the user in rating
the restaurant regarding its features. The feedback is provided to the
restaurant admin and the average feedback of the users can be viewed in
a dynamic web page which can be seen by everyone. The app also helps the
admin to send Birthday/Anniversary wishes to the users on their birthday
or anniversary. The app is developed using Eclipse and Android SDK.