Feedback of a restaurant using an Android application: (Feb 2014)
A Feedback Application for a restaurant which helps the user in rating
the restaurant regarding its features. The feedback is provided to the
restaurant admin and the average feedback of the users can be viewed in
a dynamic web page which can be seen by everyone. The app is developed
using Eclipse and Android SDK.