Multiple Object Tracking With Robust Occlusion Handling
The project is based on the field of Image Processing. In this project,
we intend to provide an approach to the problem of Multiple Object
Tracking, which has a wide range of applications such as video editing,
video surveillance, security etc. The main focus is to track multiple
objects in a scene individually in the presence of Occlusion. The
project consists of webpage through which one can upload the videos to
be tracked or Start webcam and track objects entering the webcam frame
in real time. The project was developed using MATLAB.